requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
openai
passlib
pyjwt
pytest
pillow
torch
transformers
